Refactor UI structure and enhance image caching
Updated the UI framework to improve maintainability by introducing base dialog and component classes. Fixed  image caching system for better performance when displaying card images. This change also ensures that temporary files are properly managed with revisions in the .gitignore. The overall user experience will be smoother with faster image loading and a more organized codebase.

No issues referenced.
